handle community admin removal

+++ src/main/java/taskaya/backend/DTO/communities/communityMember/responses/CommunityMemberSettingsResponseDTO.java
@@ -13,6 +13,6 @@ public class CommunityMemberSettingsResponseDTO {
     CommunityMemberCommunityProfileDTO nameAndPicture;
     String description;
     Long positionId;
-    Integer currentPercentage;
+    Double currentPercentage;
     Integer futurePercentage;
 }
